cli/requirements.txt
click==8.0.1
slack-webhook==1.0.6
python-dotenv==0.17.1
